clearing warnings
git-svn-id: svn://orca.physics.unc.edu/Orca64@11967 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Experiments/MJDTestCryoVac/ORMJDTestCryostats/ORMJDTestCryoView.h
@@ -21,6 +21,6 @@
 	int tag;
 }
 - (NSUInteger) tag;
-- (void) setTag:(int)aValue;
+- (void) setTag:(NSUInteger)aValue;
 - (void) drawString:(NSString*)aString atPoint:(NSPoint)aPoint;
 @end
